Move some methods to IngestJobStatusFromJobTestData
